+++ .github/workflows/pythonpackage.yml
@@ -40,7 +40,7 @@ jobs:
         pylint --errors-only --score=n pytest_adaptavist
         pylint --exit-zero --score=n --disable=E,R --max-line-length=160 --enable=useless-suppression pytest_adaptavist
         pylint --errors-only --score=n --load-plugins pylint_pytest tests
-        pylint --exit-zero --score=n --disable=E,R --max-line-length=160 --enable=useless-suppression --load-plugins pylint_pytest tests
+        pylint --exit-zero --score=n --disable=E,R,C0103 --max-line-length=160 --enable=useless-suppression --load-plugins pylint_pytest tests
     - name: Lint with mypy
       run: |
         python -m pip install mypy
@@ -66,4 +66,4 @@ jobs:
         pip install -e .
     - name: Test with pytest
       run: |
-        pytest --cov=pytest_adaptavist tests/unit_tests
+        pytest --cov=pytest_adaptavist tests
