+++ tests/__init__.py
@@ -1,0 +1,6 @@
+"""Unit and system tests."""
+
+from typing import Tuple
+from unittest.mock import MagicMock
+
+AdaptavistFixture = Tuple[MagicMock, MagicMock, MagicMock]
